Correçao de erros e povoamento da base de dados

+++ Projeto/Projeto/Models/acao.cs
@@ -15,13 +15,13 @@ namespace Projeto.Models
     {
         public string acaoID{ get; set; }
         public DateTime hora{ get; set; }
-        public float low{ get; set; }
-        public float high{ get; set; }
-        public float avg{ get; set; }
+        public double low{ get; set; }
+        public double high { get; set; }
+        public double avg { get; set; }
         public string empresaID{ get; set; }
 
         public Acao(){ }
-        public Acao(string aid, DateTime hora, float low, float high, float avg, string eid)
+        public Acao(string aid, DateTime hora, double low, double high, double avg, string eid)
         {
             this.acaoID = aid;
             this.hora = hora;
